VcfParser abstract class
This PR implements a generic VcfParser class and a derived class based
on Pyvcf.
This is a preparation for a second PR in which we add the support for
Nucleus based parser.
